+++ ux_valio/__init__.py
@@ -67,7 +67,7 @@ from ux_valio.validators import (
     ValueValidator,
 )
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 __all__ = [
     "AadhaarCardValidator",